vector dct_quantize_row

+++ c63.h
@@ -107,7 +107,7 @@ uint8_t get_byte(FILE *fp);
 int read_bytes(FILE *fp, void *data, unsigned int sz);
 uint16_t get_bits(struct entropy_ctx *c, uint8_t n);
 
-void dct_quant_block_8x8(int16_t *in_data, int16_t *out_data, uint8_t *quant_tbl);
+void dct_quant_block_8x8(int *in_data, int16_t *out_data, uint8_t *quant_tbl);
 void dequant_idct_block_8x8(int16_t *in_data, int16_t *out_data, uint8_t *quant_tbl);
 void sad_block_8x8(uint8_t *block1, uint8_t *block2, int stride, int *result);
